fix: request_tool actually performs catalog lookup for BROKEN/IMPROVE kinds (fixes #343 review)
- Extracted buildClientToolCatalog helper into packages/shared-utils/src/mcp-catalog.ts
  (shared across request_tool + tool-request-dedupe); also moved discoverMcpServer and
  normalizeUrl there so copilot-api/mcp-discovery.ts becomes a thin re-export shim
- request_tool: when kind!=NEW_TOOL, calls buildClientToolCatalog(db, clientId,
  { encryptionKey }) and warns only when normalizedName is absent from the catalog
- Dropped result.isNew gate — row-newness is orthogonal to catalog membership
- tool-request-dedupe: replaced ~80 lines of inline catalog-building with
  buildClientToolCatalog call (passes warnings array for backward-compat DedupeResult)
- findMany in tool-request-dedupe uses include (all fields), so kind: true was not needed
