+++ packages/shared-utils/src/index.ts
@@ -36,6 +36,17 @@ export { createToolRequestGithubIssue, buildToolRequestIssueBody, ToolRequestNot
 export type { CreateGithubIssueInput, CreateGithubIssueResult } from './tool-request-github.js';
 export { withTicketLock } from './advisory-lock.js';
 export type { PrismaTx } from './advisory-lock.js';
+export {
+  normalizeUrl,
+  discoverMcpServer,
+  buildClientToolCatalog,
+} from './mcp-catalog.js';
+export type {
+  McpDiscoveryConfig,
+  McpToolInfo,
+  McpDiscoveryResult,
+  ClientCatalogOpts,
+} from './mcp-catalog.js';
 export { isTransientApiError } from './transient-error.js';
 export {
   initEmptyKnowledgeDoc,
